베팅 금액을 Double로 수정

+++ src/main/kotlin/blackjack/common/PlayerProperty.kt
@@ -1,3 +1,3 @@
 package blackjack.common
 
-class PlayerProperty(val name: String, val bet: Int)
+class PlayerProperty(val name: String, val bet: Double)
