+++ src/main/kotlin/blackjack/domain/player/Player.kt
@@ -4,7 +4,7 @@ import blackjack.domain.bet.Bet
 import blackjack.domain.card.Card
 import blackjack.domain.card.Cards
 
-open class Player(val name: String, bet: Int, startingCards: List<Card>) {
+open class Player(val name: String, bet: Double, startingCards: List<Card>) {
     val cards: Cards = Cards(startingCards)
 
     val bet: Bet = Bet(bet)
